IBN-535 remove old state listeners for arithmetic
If an arithmetic function changes the logic link for example from
'and' to 'or', old statechangelistener remain in the added functions.
These listeners are added to notify the logic of state changes in any
added function which require the logic itself to revaluate, if itself
has to change the state. By removing the old statechangelisteners, new
ones will be added, which represent the correct logic link, when the
logic link changes.