IBN-2008: Threshold: Add optional factor parameter
When handling a child whose state should be interpreted as PercentType
(i.e. whose state is in [0..1] but represents a percentage in [0..100]),
we need to account for this factor when calculating the state of our function.

Since we cannot access the function type of the child item in the calculation,
we instead pass a factor as the optional fifth parameter to the function,
which it can then multiply the child item state with,
and which defaults to one.