IBN-2098: Pin gulp-inject version
The current version range allows for either gulp-inject 4.3.2 or 4.3.3 to be installed.
Version 4.3.2 depends on ansi-color 1.1.0, which is fine,
while version 4.3.3 depends on ansi-color 3, which requires node 6 and breaks the build.
Upgrading node itself leads to other seemingly unrelated errors
(failure to git clone some github wikis during npm install, but only on jenkins)
which makes this not an option.

This change pins gulp-inject to the known working version.