IBN-613: Add temporary storage to preserve sd card life
The PersistentInbox used the mapbd storage to store all discovery results,
which put wear on the sd card with the writes to it. To prolong the sd card
lifetime, the storage for inbox entries should be in ram. The PersistentInbox
is left, even tho the name is kinda missleading, because we want to obtain as
much compability with esh. A second inbox, which is used through higher order
ranking doesn't work either, because the Inbox itself registers as a Discovery-
Listener. Thus the inbox would be doubled. Through an additional storage
implementation, copied from the test bundle into a new bundle, we can now use
this TemporaryStorage in place of the mapdb storage.

To actually use this new storage implementation in PersistentInbox,
add the following to your configuration:

    org.eclipse.smarthome.config.discovery.internal.PersistentInbox:StorageService.target=(type=temporaryStorage)